+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: Adding objects to a collection should be allowed unless locked by as user
 
 1.1.1 (2022-06-09)
 ==================